% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_handlers.R
\name{add_empty_wells}
\alias{add_empty_wells}
\title{Add empty wells ### DOCUMENTATION NOT FINISHED######}
\usage{
add_empty_wells(layout, plate_type = "384", .well_col = "well", ...)
}
\arguments{
\item{layout}{A layout tibble}

\item{plate_type}{the type of plate}

\item{.well_col}{A string--the name of the column that holds the well}

\item{...}{Unused, for extensibility}
}
\value{
the same layout, but with all wells in the chosen plate type
}
\description{
Add wells missing from the input layout. These wells are assigned \code{NA} for all variables.
}
