% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_handlers.R
\name{add_missing_wells}
\alias{add_missing_wells}
\title{Add wells not included in a layout}
\usage{
add_missing_wells(layout, plate_type = "384", .well_col = "well", ...)
}
\arguments{
\item{layout}{A layout tibble}

\item{plate_type}{The type of plate to make well names for; can be
\itemize{
\item \code{96}, for 96 well plates.
\item \code{384}, for 384 well plates.
\item "custom", for a plate containing user-specified rows and columns to \code{plate_rows} and \code{plate_cols}, which are passed to \code{\link[=dimensions_to_wells]{dimensions_to_wells()}}.
}}

\item{.well_col}{A string--the name of the column that holds the well}

\item{...}{Unused, for extensibility.}
}
\value{
the same layout, but with all wells in the chosen plate type
}
\description{
If a layout does not contain every well for a given plate type (e.g. 384),
add the missing wells. Assign "NA" for all experimental variables.
}
