% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_handlers.R
\name{extract_plate_row}
\alias{extract_plate_row}
\alias{extract_plate_col}
\title{Extract plate dimensions}
\usage{
extract_plate_row(well_names)

extract_plate_col(well_names)
}
\arguments{
\item{well_names}{A character vector containing well names, e.g. c("A1", "A2", "B1", "B2").}
}
\description{
Extract the suspected plate row or column from a well name, or vector of well names.
\itemize{
\item \code{extract_plate_row()} returns suspected rows, for example, "A", from "A1". Specifically, returns the character element of the supplied well name(s).
\item \code{extract_plate_col()}, returns suspected columns, for example, "1", from "A1". Specifically, returns the numeric element of the supplied well name(s).
}
}
\section{Functions}{
\itemize{
\item \code{extract_plate_row()}: Extract the plate row from a well name

\item \code{extract_plate_col()}: Extract the plate column from a well name

}}
\examples{
# extract the suspected plate row
extract_plate_row(c("A1", "A2", "B1", "B2"))
# [1] "A" "A" "B" "B"

# extract the suspected plate column
extract_plate_col(c("A1", "A2", "B1", "B2"))
# [1] 1 2 1 2

}
\concept{plate helpers}
