% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{platedescribr_example}
\alias{platedescribr_example}
\title{Get path to platedefinr examples}
\usage{
platedescribr_example(path = NULL)
}
\arguments{
\item{path}{Name of file.}
}
\value{
Path to the supplied file.
}
\description{
platedefinr comes with a number of sample files in its 'inst/extdata' directory. Use \code{platedescribr_example()} to retrieve the path to an available example.
example.
}
\details{
This documentation is minimally modified from the analogous file in the vroom package. (see https://github.com/tidyverse/vroom/blob/main/R/example.R)
}
\examples{
# Get path to one example
# platedescribr_example("dose_response.csv")

}
