% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_layout.R
\name{read_layout}
\alias{read_layout}
\title{Read a plate layout file into a tibble}
\usage{
read_layout(filepath, ...)
}
\arguments{
\item{filepath}{Path to a file containing a plate layout. Accepts .csv, .txt, .xlsx, and .xls.}

\item{...}{Unused, for extensibility.}
}
\value{
Returns a tibble with each well position as row, and the user-provided experimental variables in the following columns. All outputs contain the following columns: row, column, well, and condition. The condition column contains all experimental variables in a single, underscore-separated string. If all experimental variables are defined in the layout, wells with identical entries in the "condition" column are technical replicates.
}
\description{
read_layout() reads a raw plate layout file (.csv, .txt, .xls, or .xlsx), and returns it as a formatted tibble. Variables are coerced to numeric if indicated by readr::parse_guess(). The originally required "Type" column heading is now optional.
}
