% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_plate.R
\name{summarise_plate}
\alias{summarise_plate}
\title{Print a summary of plate contents}
\usage{
summarise_plate(layout, plate_notes = "none", ...)
}
\arguments{
\item{layout}{A tibble of a tidied layout #### UPDATE THIS DOCUMENTATION}

\item{plate_notes}{A list containing notes, as strings, to be included in the summary.}

\item{...}{Unused, for extensibility.}
}
\value{
A printable summary of the contents of the plate, including
\itemize{
\item The number of wells
\item The number of unique experimental conditions
\item The name of each experimental variable
\item The value(s) for each experimental variable, and the number of wells in which that value occurs.
\item Any user-specified notes on the experiment.
}
}
\description{
Given a layout, create a printed summary of layout contents.
}
