% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_plate.R
\name{summarise_variable}
\alias{summarise_variable}
\title{Print a summary of a variable's contents}
\usage{
summarise_variable(layout, variable_name, ...)
}
\arguments{
\item{layout}{A tibble of a tidied layout #### UPDATE THIS DOCUMENTATION}

\item{variable_name}{The name of the column containing the experimental
variable to be summarised, as a string.}

\item{...}{Unused, for extensibility.}
}
\value{
A glued string stating the variable name, and the number of wells in
which each value of that variable occurs.
}
\description{
Given a layout, prints the name of a given variable, and the number of wells
in which each of its values appears. Used as a part of \code{\link[=summarise_plate]{summarise_plate()}}.
}
