% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_handlers.R
\name{well_names}
\alias{well_names}
\title{Get the names of wells from a given plate  ### DOCUMENTATION NOT FINISHED######}
\usage{
well_names(
  platetype,
  ...,
  dim1 = NULL,
  dim2 = NULL,
  .leading_zeroes = FALSE,
  .as_factor = FALSE
)
}
\arguments{
\item{platetype}{type of plate}

\item{...}{Unused, for extensibility}

\item{dim1}{plate rows}

\item{dim2}{plate cols}

\item{.leading_zeroes}{add leading zeroes?}

\item{.as_factor}{make the output a factor?}
}
\value{
a vector of well names.
}
\description{
Get the names of wells from a given plate  ### DOCUMENTATION NOT FINISHED######
}
\examples{
well_names(platetype = 384)

}
