% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_handlers.R
\name{wells_to_cols}
\alias{wells_to_cols}
\title{Wells to cols ## DOCUMENTATION NOT FINISHED######}
\usage{
wells_to_cols(well_names)
}
\arguments{
\item{well_names}{A vector containing only the character elements of each well name.}
}
\value{
A vector containing only the numeric elements of each well name.
}
\description{
Given a well name (e.g. "A1"), extract the plate column (1).
}
\examples{
wells_to_cols(c("A1", "A2", "B1", "B2"))
# [1] 1 2 1 2

}
