% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_handlers.R
\name{wells_to_rows}
\alias{wells_to_rows}
\title{Wells to rows ## DOCUMENTATION NOT FINISHED######}
\usage{
wells_to_rows(well_names)
}
\arguments{
\item{well_names}{A vector of well names.}
}
\value{
A vector containing only the character elements of each well name.
}
\description{
Given a well name (e.g. "A1"), extract the plate row ("A").
}
\examples{
wells_to_rows(c("A1", "A2", "B1", "B2"))
# [1] "A" "A" "B" "B"

}
