% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_handlers.R
\name{expand_wells}
\alias{expand_wells}
\title{Expand wells}
\usage{
expand_wells(dim1, dim2, .leading_zeroes = FALSE)
}
\arguments{
\item{dim1}{first dimension to expand}

\item{dim2}{second dimension to expand}

\item{.leading_zeroes}{whether or not to add leading zeroes}
}
\value{
vector of well names
}
\description{
Expand wells
}
